pyproject.toml
[tool.poetry]
name = "etherfetch"
version = "0.0.1"
description = "Fetch verified contracts from etherscan and assemble them locally such that they can be compiled with solc binary"
authors = ["TilakMaddy <tilak.madichetti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
typer = {extras = ["all"], version = "^0.12.3"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
